fix: update Makefile to trigger full re-compiling on .h file changes
Change the Makefile to ensure that any changes to .h header files will
trigger a full recompilation of the program/file object.  Previously,
modifications to .h files did not lead to a complete rebuild, which
could result in out-of-date binaries and unexpected behavior.

With this update, the build process is now correctly dependent on all
relevant header files, ensuring that any changes propagate through to
the final executable/file object.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>