refactor ipv6 acc routing to separate protocol components
The ipv6_route() function has been refactored by splitting it into two
distinct components.

The ipv6 routing part focuses on the dependent protocol aspects, while
the second part addresses the independent protocol functionality,
specifically where the FIB lookup occurs through the fib_lookup(...)
function.
This separation of "concerns" not only enhances the clarity of the code
but also promotes code reuse. By isolating the independent protocol
part, the same logic can be leveraged to implement routing for other
protocols, such as IPv4.

*** Please note that routing acceleration in eBPF/XDP mode is disabled
by default. ***
This is because integrating this solution with iptables/ip6tables for
setting different traffic classes in SRv6 packets is not feasible. When
routing acceleration is enabled, the traditional kernel stack is
bypassed, and packets are sent directly to the device responsible for
reaching the next hop after the lookup process.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>