remove unnecessary vlan_check()
The vlan_check() function has been deemed unnecessary in our current
implementation. When parsing the Ethernet header, we specifically
request the protocols carried by Ethernet, namely IPv4 and IPv6. If the
Ethernet protocol does not match either of these, the packet is passed
directly to the kernel stack for further processing.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>