introduce new priority-based trace pipe printing API
This commit introduces a new set of APIs for printing messages on the trace
pipe, allowing differentiation of print behavior based on assigned
priorities.  The priorities range from EMERG (0) to DEBUG (7), with lower
integer values representing higher priority messages.

The default PRINT_LEVEL is set to WARNING, ensuring that only messages
of this priority or higher are displayed.

Additionally, debug and critical messages have been integrated into the
encap and decap XDP programs to enhance visibility into the processing
logic.  This improvement will facilitate better debugging and monitoring of
the packet processing flow.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>