add support for table ID in XDP routing acceleration
This patch  introduces a new member called "tbid" to the fib_res_lookup
structure, which allows for the specification of the table used during
the lookup operation (using the XDP routing acceleration facility).

To perform a lookup on a specified table, the flags
"BPF_FIB_LOOKUP_DIRECT | BPF_FIB_LOOKUP_TBID" must be also set in the
"flags" member of fib_res_lookup.
This enhancement improves the flexibility and functionality of the XDP
routing acceleration mechanism.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>