fix: adjust MTU Settings/Implement XDP workarounds for VDC Networking
This commit change the MTU configuration for interfaces within the
virtual data center network, setting it to 2048. This value has been
chosen to comply with XDP constraints, as XDP programs cannot be
attached to veth pairs with an MTU exceeding 4096 (the size of a page
frame). While we have decided on 2048 for the time being, there is room
for slight increases without surpassing safe limits.

In addition, two workarounds have been implemented to ensure the proper
operation of XDP programs within the virtual data center networking:

 1. Checksum Offload Disabled: We have disabled checksum offloading for
    both transmit and receive operations on veth pairs to prevent packet
    drop during reception, which was causing issues previously.

 2. eBPF/XDP "fib_lookup" neigh: The "fib_lookup" helper function in
    eBPF/XDP does not populate the neighbor cache. As a result, when
    packets are decapsulated and sent back to the kernel stack, the routing
    lookup defaults to the "main" table, as the packets do not originate
    from any VRF or its associated slaves (note that IPv4 routes for a
    specific customer are managed by a VRF).

    To address this, we have configured a route in the main routing
    table to direct decapsulated IPv4 traffic for a customer to its
    corresponding VRF. This route is only necessary for the initial
    decapsulated IPv4 packet, as the neighbor cache will automatically
    populate for subsequent packets, allowing them to be seamlessly
    forwarded using the eBPF/XDP forwarding mechanism.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>